requirements.txt
click==8.1.7
numpy==2.1.2
opencv-python==4.10.0.84
platformdirs==4.3.6
scenedetect==0.6.4
tqdm==4.66.5
openai
langchain
langchain-community
decord
transformers
einops
timm
accelerate>=0.26.0
sentencepiece
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
protobuf==4.25.5
Pygments==2.18.0
requests==2.32.3
rich==13.9.2
scenedetect==0.6.4
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
decord
transformers
einops
timm
accelerate>=0.26.0
sentencepiece